move `App` to `app` folder
