+++ src/app/main.ts
@@ -1,7 +1,7 @@
 import '../public/resources/css/app.css';
 import '@fortawesome/fontawesome-free/css/all.css';
 
-import App from './models/App';
+import App from './App';
 import loadComponents from './helpers/loadComponents';
 import getHTMLElement from './helpers/getHTMLElement';
